chore: add ThinkingSegment
