+++ src/components/Assistant/ChatMessage.tsx
@@ -1,5 +1,4 @@
-import { Brain, ChevronDown, ChevronUp } from "lucide-react";
-import { useState, memo } from "react";
+import { memo } from "react";
 import { useTranslation } from "react-i18next";
 
 import type { Message } from "./types";
@@ -7,6 +6,7 @@ import Markdown from "./Markdown";
 import { formatThinkingMessage } from "@/utils/index";
 import logoImg from "@/assets/icon.svg";
 import { SourceResult } from "./SourceResult";
+import { ThinkingSegment } from "./ThinkingSegment";
 // import { ThinkingSteps } from "./ThinkingSteps";
 
 interface ChatMessageProps {
@@ -33,7 +33,6 @@ export const ChatMessage = memo(function ChatMessage({
   isThinkTyping,
 }: ChatMessageProps) {
   const { t } = useTranslation();
-  const [isThinkingExpanded, setIsThinkingExpanded] = useState(true);
 
   const isAssistant = message._source?.type === "assistant";
   const messageContent = message._source?.message || "";
@@ -67,50 +66,11 @@ export const ChatMessage = memo(function ChatMessage({
                 type={segment.sourceType}
               />
             ) : segment.isThinking ? (
-              <div className="space-y-2 mb-3 w-full">
-                <button
-                  onClick={() => setIsThinkingExpanded((prev) => !prev)}
-                  className="inline-flex items-center gap-2 px-2 py-1 rounded-xl transition-colors border border-[#E6E6E6] dark:border-[#272626]"
-                >
-                  {isThinkTyping ? (
-                    <>
-                      <Brain className="w-4 h-4 animate-pulse text-[#999999]" />
-                      <span className="text-xs text-[#999999] italic">
-                        {t(`assistant.message.steps.${segment.sourceType}`)}
-                      </span>
-                    </>
-                  ) : (
-                    <>
-                      <Brain className="w-4 h-4 text-[#999999]" />
-                      <span className="text-xs text-[#999999]">
-                        {segment.sourceType === "think"
-                          ? t("assistant.message.steps.thoughtTime")
-                          : t(`assistant.message.steps.${segment.sourceType}`)}
-                      </span>
-                    </>
-                  )}
-                  {segment.thinkContent &&
-                    (isThinkingExpanded ? (
-                      <ChevronUp className="w-4 h-4" />
-                    ) : (
-                      <ChevronDown className="w-4 h-4" />
-                    ))}
-                </button>
-                {isThinkingExpanded && segment.thinkContent && (
-                  <div className="pl-2 border-l-2 border-[e5e5e5]">
-                    <div className="text-[#8b8b8b] dark:text-[#a6a6a6] space-y-2">
-                      {segment.thinkContent.split("\n").map(
-                        (paragraph, idx) =>
-                          paragraph.trim() && (
-                            <p key={idx} className="text-sm">
-                              {paragraph}
-                            </p>
-                          )
-                      )}
-                    </div>
-                  </div>
-                )}
-              </div>
+              <ThinkingSegment
+                sourceType={segment.sourceType || ""}
+                thinkContent={segment.thinkContent || ""}
+                isThinkTyping={isThinkTyping}
+              />
             ) : segment.text ? (
               <div className="space-y-4">
                 <Markdown
